Adds a format spec for truncating strings and appending a suffix at the end.
The motivation behind this change is that sometimes when artist's names or titles get
truncated is impossible to know if the string you are seeing is the actual name or
if it has been shortened and you are missing information.

This could have been emulated to some degree with the already existing spec:

"{title:…<10.9}" which would do the next:

hello world → hello wor…
hello       → hello……………

In the second case is clear that the string has not been shortened, but the excessive
filler at the end seems wrong.

Also, this built-in function only supports filling up with one char, the feature added
supports multi-char suffixes (although they won't repeat to fill up excess space).

Unit tests added and documentation updated.